Add cli_utils.py
