+++ src/qfed/utils.py
@@ -1,13 +1,10 @@
-'''
-This module contains:
-    - Engines that provide uniform interface for acessing data
-      saved in commonly used formats such as HDF, NetCDF4, etc. 
-    - Functions that are used in the L3A and L3B scripts/CLI.
-'''
+"""
+This module contains readers for retrieving data and attributes 
+from files using formats such as HDF, NetCDF4, etc. 
+"""
 
 import sys
 import os
-import yaml
 import logging
 import abc
 
@@ -21,55 +18,15 @@ else:
     ABC = abc.ABCMeta('ABC', (), {})
 
 
-def read_config(config):
-    """
-    Parses the QFED config file into a dictionary.
-    """
-    with open(config) as file:
-        try:
-            data = yaml.safe_load(file)
-        except yaml.YAMLError as err:
-            data = None
-            logging.critical(err)
-
-    return data
-
-
-def get_path(path, timestamp=None):
-    """
-    Generate an optionally timestamped path.
-    """
-    if isinstance(path, list):
-        result = path
-    else:
-        result = [path]
-
-    if timestamp is not None:
-        result = [atom.format(timestamp) for atom in result]
-
-    return os.path.join(*result)
-
-
-def display_description(version, text):
-    """
-    Displays the QFED version and a text.
-    """
-    logging.info('')
-    logging.info(f'QFED {version}')
-    logging.info('')
-    logging.info(text)
-    logging.info('')
-
-
 class DatasetAccessEngine(ABC):
     """
     Abstract engine to access data.
     """
 
-    def __init__(self, msg="Could not open file <{file}> - ignoring it."):
+    def __init__(self, msg="Could not open file '{file}' - ignoring it."):
         self._msg = msg
 
-    def _message_on_file_error(self, file):
+    def _message_on_open_file_error(self, file):
         logging.warning(self._msg.format(file=file))
 
     @abc.abstractmethod
@@ -94,7 +51,7 @@ class DatasetAccessEngine_HDF4(DatasetAccessEngine):
         try:
             f = SD.SD(file)
         except SD.HDF4Error:
-            self._message_on_file_error(file)
+            self._message_on_open_file_error(file)
             f = None
 
         return f
@@ -133,7 +90,7 @@ class DatasetAccessEngine_NetCDF4(DatasetAccessEngine):
         try:
             f = nc.Dataset(file)
         except IOError:
-            self._message_on_file_error(file)
+            self._message_on_open_file_error(file)
             f = None
 
         return f
